Code (v3.1)
my partner and I have decided to do a complet revamp of our project. we noticed our languagelearning class was getting too long so we decided to make new classes and put different methods in each, doing this we have shorten out language learning class and made it easier for people that are seeing the code for the first time be able to understand the code easily.